Subsumes the full i686→x86_64 port plus a long tail of follow-up
work: 4-level paging, higher-half + HHDM, per-process PML4,
FXSAVE/FXRSTOR context switching, SYSCALL/SYSRET fast path, real
dynamic linking (ld-lighthos.so.1 + libulib.so.1 / libvibc.so.1 /
dlopen iface), signals with user handlers, pgroups + terminal
foreground + Ctrl-C/Ctrl-Z/fg/bg, chroot, pipes, strace ring,
alarm, mount/umount, mmap_anon, FAT32 r/w + LFN, the shell's line
editor (history, tab complete, word nav, Ctrl-A/E/U/K), real Lua
(dynamic, require() with pure-Lua AND C modules via dlopen, os.*
and io.popen), vi, hexdump/pagemap/regions, SYS_PEEK/PAGEMAP/REGIONS,
terminal window-size probe + cache, raw-mode ioctl, VGA+PS/2
console routing, VGA mode-13h graphics + flappy bird, save/restore
font across graphics round-trips, production vs test binary split,
27 test scripts / 60 assertions all green.

See individual commit messages for the per-feature story.

+++ src/kernel/timer.h
@@ -3,7 +3,7 @@
 
 #include "include/types.h"
 
-void timer_init(uint32_t frequency_hz);
-uint32_t timer_get_ticks(void);
+void     timer_init(uint32_t frequency_hz);
+uint64_t timer_get_ticks(void);
 
 #endif
