+++ src/mm/heap.h
@@ -3,11 +3,11 @@
 
 #include "include/types.h"
 
-void  heap_init(uint32_t start, uint32_t size);
-void *kmalloc(uint32_t size);
+void  heap_init(uint64_t start, uint64_t size);
+void *kmalloc(uint64_t size);
 void  kfree(void *ptr);
 
-uint32_t heap_get_used(void);
-uint32_t heap_get_free(void);
+uint64_t heap_get_used(void);
+uint64_t heap_get_free(void);
 
 #endif
